+++ src/app/core/pages/Shared/index.js
@@ -1,0 +1,1 @@
+export { default } from './Shared'
